change project java version to 11

+++ src/main/java/com/example/controller/UserController.java
@@ -12,11 +12,14 @@ import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 /**
@@ -89,12 +92,24 @@ public class UserController {
         else return new ResponseEntity<>("Error", HttpStatus.BAD_REQUEST);
     }
 
-    @Operation(responses = @ApiResponse(responseCode = "200", content = @Content(schema = @Schema(hidden = true))))
+    @Operation(responses = @ApiResponse(responseCode = "200", content = @Content(schema = @Schema(hidden = true))),
+            security = {@SecurityRequirement(name = "Authorization")})
     @DeleteMapping(value = "deleteUser")
     public ResponseEntity<?>deleteUser(@RequestParam int id){
-        if (userServices.deleteUser(id)){
-            return new ResponseEntity<>("user is delete", HttpStatus.OK);
-        }
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        if (authentication != null && authentication.getAuthorities().stream().anyMatch(a -> a.getAuthority().equals("USER")))
+        {
+            if (userServices.deleteUser(id)){
+                return new ResponseEntity<>("user is delete", HttpStatus.OK);
+            }
             else return new ResponseEntity<>("Error", HttpStatus.BAD_REQUEST);
+        }
+        else{
+            if (authentication == null){
+                return new ResponseEntity<>("Please login", HttpStatus.UNAUTHORIZED);
+            }
+            else
+                return new ResponseEntity<>("You don't have permission", HttpStatus.UNAUTHORIZED);
+        }
     }
 }
